smearHadesMomentum: declare the argument 'p' as 'real(8)' * in order to be compatible with Python/numpy * avoids errors of type "ValueError: failed to initialize intent(inout) array -- expected elsize=4 but got 8"

+++ readHAFT2.f90
@@ -849,7 +849,7 @@ contains
 
 
   subroutine smearHadesMomentum(p,mode,pid)
-      real, intent(inout) :: p(0:3)
+      real(8), intent(inout) :: p(0:3)
       integer, intent(in) :: mode, pid
 
       real :: mom4(4)
